+++ components/breadcrumb/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @westpac/breadcrumb
 
+## 2.2.0
+
+### Minor Changes
+
+- 20f9f31f: Proptypes created using typescript and type fixes
+
+### Patch Changes
+
+- Updated dependencies [bf4dc0e5]
+- Updated dependencies [20f9f31f]
+  - @westpac/core@3.2.0
+  - @westpac/icon@2.1.0
+
 ## 2.1.0
 
 ### Minor Changes