+++ components/panel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @westpac/panel
 
+## 2.2.0
+
+### Minor Changes
+
+- 20f9f31f: Proptypes created using typescript and type fixes
+
+### Patch Changes
+
+- Updated dependencies [bf4dc0e5]
+- Updated dependencies [20f9f31f]
+  - @westpac/core@3.2.0
+  - @westpac/heading@2.2.0
+  - @westpac/body@2.2.0
+
 ## 2.1.0
 
 ### Minor Changes